requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Data Science & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.2.0

# Progress Bar
tqdm>=4.65.0

# Grad-CAM for Explainability
pytorch-grad-cam>=1.4.8

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.22.0
